# Change query plan indent size