@yfuruyama 2020-07-16
@apstndb What do you think about this change?